Add `ModuleSource::Block`
